+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to the project will be documented in this file.
 
 - ![breaking] Support PostCSS v8
 - ![breaking] PostCSS is no more a `peerDependency`
+- ![breaking] Source Map accept the following option: `true`, `false` and `'inline'` values
 - A new option, `syntax`, is available in the configuration. [Read more](https://github.com/postcss/postcss#syntaxes) on the PostCSS documentation.
 
 **Dependencies:**
@@ -85,7 +86,7 @@ All notable changes to the project will be documented in this file.
 
 **Features:**
 
-- Source Map option accept `true`, `false` and `'external'` values
+- Source Map accept the following option: `true`, `false` and `'external'` values
 - Source Map option is disabled by default
 
 **Fixes:**